Trixi.jl API

Trixi.Trixi — Module

Trixi

Trixi.jl is a numerical simulation framework for hyperbolic conservation laws. A key objective for the framework is to be useful to both scientists and students. Therefore, next to having an extensible design with a fast implementation, Trixi is focused on being easy to use for new or inexperienced users, including the installation and postprocessing procedures.

To get started, run your first simulation with Trixi using

trixi_include(default_example())

See also: trixi-framework/Trixi.jl

Trixi.boundary_condition_periodic — Constant

boundary_condition_periodic = BoundaryConditionPeriodic()

A singleton struct indicating periodic boundary conditions.

Trixi.AMRCallback — Type

AMRCallback(semi, controller [,adaptor=AdaptorAMR(semi)];
            interval,
            adapt_initial_condition=true,
            adapt_initial_condition_only_refine=true)

Performs adaptive mesh refinement (AMR) every interval time steps for a given semidiscretization semi using the chosen controller.

Trixi.AliveCallback — Type

AliveCallback(analysis_interval=0, alive_interval=analysis_interval÷10)

Inexpensive callback showing that a simulation is still running by printing some information such as the current time to the screen every alive_interval time steps. If analysis_interval ≂̸ 0, the output is omitted every analysis_interval time steps.

Trixi.AnalysisCallback — Type

AnalysisCallback(semi; interval=0,
                       save_analysis=false,
                       output_directory="out",
                       analysis_filename="analysis.dat",
                       extra_analysis_errors=Symbol[],
                       extra_analysis_integrals=())

Analyze a numerical solution every interval time steps and print the results to the screen. If save_analysis, the results are also saved in joinpath(output_directory, analysis_filename).

Additional errors can be computed, e.g. by passing extra_analysis_errors = [:primitive].

Further scalar functions func in extra_analysis_integrals are applied to the numerical solution and integrated over the computational domain. See Trixi.analyze, Trixi.pretty_form_utf, Trixi.pretty_form_ascii for further information on how to create custom analysis quantities.

Trixi.CarpenterKennedy2N43 — Type

  CarpenterKennedy2N43()

Carpenter, Kennedy (1994) Third order 2N storage RK schemes with error control

Trixi.CarpenterKennedy2N54 — Type

CarpenterKennedy2N54()

The following structures and methods provide a minimal implementation of the low-storage explicit Runge-Kutta method of

Carpenter, Kennedy (1994) Fourth order 2N storage RK schemes, Solution 3

using the same interface as OrdinaryDiffEq.jl.

Trixi.CompressibleEulerEquations1D — Type

CompressibleEulerEquations1D

The compressible Euler equations for an ideal gas in one space dimension.

Trixi.CompressibleEulerEquations2D — Type

CompressibleEulerEquations2D

The compressible Euler equations for an ideal gas in two space dimensions.

Trixi.CompressibleEulerEquations3D — Type

CompressibleEulerEquations3D

The compressible Euler equations for an ideal gas in three space dimensions.

Trixi.ControllerThreeLevel — Type

ControllerThreeLevel(semi, indicator; base_level=1,
                                      med_level=base_level, med_threshold=0.0,
                                      max_level=base_level, max_threshold=1.0)

An AMR controller based on three levels (in decending order of precedence):

set the target level to max_level if indicator > max_threshold

set the target level to med_level if indicator > med_threshold; if med_level < 0, set the target level to the current level

set the target level to base_level otherwise

Trixi.ControllerThreeLevelCombined — Type

ControllerThreeLevelCombined(semi, indicator_primary, indicator_secondary;
                             base_level=1,
                             med_level=base_level, med_threshold=0.0,
                             max_level=base_level, max_threshold=1.0,
                             max_threshold_secondary=1.0)

An AMR controller based on three levels (in decending order of precedence):

set the target level to max_level if indicator_primary > max_threshold

set the target level to med_level if indicator_primary > med_threshold; if med_level < 0, set the target level to the current level

set the target level to base_level otherwise

If indicator_secondary >= max_threshold_secondary, set the target level to max_level.

Trixi.DG — Type

DG(basis, basis, surface_flux, volume_integral)

Create a discontinuous Galerkin method. If basis isa LobattoLegendreBasis, this creates a DGSEM.

Trixi.DGSEM — Type

DGSEM([RealT=Float64,] polydeg::Integer,
      surface_flux=flux_central,
      volume_integral::AbstractVolumeIntegral=VolumeIntegralWeakForm(),
      mortar=MortarL2(basis))

Create a discontinuous Galerkin spectral element method (DGSEM) using a LobattoLegendreBasis with polynomials of degree polydeg.

Trixi.HypDiffN3Erk3Sstar52 — Type

HypDiffN3Erk3Sstar52()

Five stage, second-order acurate explicit Runge-Kutta scheme with stability region optimized for the hyperbolic diffusion equation with LLF flux and polynomials of degree polydeg=3.

Trixi.HyperbolicDiffusionEquations2D — Type

HyperbolicDiffusionEquations2D

The linear hyperbolic diffusion equations in two space dimensions. A description of this system can be found in Sec. 2.5 of the book "I Do Like CFD, Too: Vol 1". The book is freely available at http://www.cfdbooks.com/ and further analysis can be found in the paper by Nishikawa DOI: 10.1016/j.jcp.2007.07.029

Trixi.HyperbolicDiffusionEquations3D — Type

HyperbolicDiffusionEquations3D

The linear hyperbolic diffusion equations in three space dimensions. A description of this system can be found in Sec. 2.5 of the book "I Do Like CFD, Too: Vol 1". The book is freely available at http://www.cfdbooks.com/ and further analysis can be found in the paper by Nishikawa DOI: 10.1016/j.jcp.2007.07.029

Trixi.IdealGlmMhdEquations2D — Type

IdealGlmMhdEquations2D

The ideal compressible GLM-MHD equations in two space dimensions.

Trixi.IdealGlmMhdEquations3D — Type

IdealGlmMhdEquations3D

The ideal compressible GLM-MHD equations in three space dimensions.

Trixi.IndicatorHennemannGassner — Type

IndicatorHennemannGassner

Indicator used for shock-capturing or AMR used by

Hennemann, Gassner (2020) "A provably entropy stable subcell shock capturing approach for high order split form DG" arXiv: 2008.12044

Trixi.IndicatorLöhner — Type

IndicatorLöhner (equivalent to IndicatorLoehner)

AMR indicator adapted from a FEM indicator by Löhner (1987), also used in the FLASH code as standard AMR indicator. The indicator estimates a weighted second derivative of a specified variable locally.

Löhner (1987) "An adaptive finite element scheme for transient problems in CFD" doi: 10.1016/0045-7825(87)90098-3

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node59.html#SECTION05163100000000000000

Trixi.LinearScalarAdvectionEquation1D — Type

LinearScalarAdvectionEquation1D

The linear scalar advection equation

\[\partial_t u + a \partial_1 u = 0\]

in one space dimension with constant velocity a.

Trixi.LinearScalarAdvectionEquation2D — Type

LinearScalarAdvectionEquation2D

The linear scalar advection equation

\[\partial_t u + a_1 \partial_1 u + a_2 \partial_2 u = 0\]

in two space dimensions with constant velocity a.

Trixi.LinearScalarAdvectionEquation3D — Type

LinearScalarAdvectionEquation3D

The linear scalar advection equation

\[\partial_t u + a_1 \partial_1 u + a_2 \partial_2 u + a_3 \partial_3 u = 0\]

in three space dimensions with constant velocity a.

Trixi.LobattoLegendreBasis — Type

LobattoLegendreBasis([RealT=Float64,] polydeg::Integer)

Create a nodal Lobatto-Legendre basis for polynomials of degree polydeg.

Trixi.ParametersEulerGravity — Type

ParametersEulerGravity(; background_density=0.0,
                         gravitational_constant=1.0,
                         cfl=1.0,
                         resid_tol=1.0e-4,
                         n_iterations_max=10^4,
                         timestep_gravity=timestep_gravity_erk52_3Sstar!)

Set up parameters for the gravitational part of a SemidiscretizationEulerGravity.

Trixi.ParsaniKetchesonDeconinck3Sstar32 — Type

ParsaniKetchesonDeconinck3Sstar32()

Parsani, Ketcheson, Deconinck (2013) Optimized explicit RK schemes for the spectral difference method applied to wave propagation problems DOI: 10.1137/120885899

Trixi.ParsaniKetchesonDeconinck3Sstar94 — Type

ParsaniKetchesonDeconinck3Sstar94()

Parsani, Ketcheson, Deconinck (2013) Optimized explicit RK schemes for the spectral difference method applied to wave propagation problems DOI: 10.1137/120885899

Trixi.PerformanceCounter — Type

PerformanceCounter

A PerformanceCounter be used to track the runtime performance of some calls. Add a new runtime measurement via put!(counter, runtime) and get the averaged runtime of all measurements added so far via take!(counter), resetting the counter.

Trixi.PositivityPreservingLimiterZhangShu — Type

PositivityPreservingLimiterZhangShu(; threshold, variables)

The fully-discrete positivity-preserving limiter of

Zhang, Shu (2011) Maximum-principle-satisfying and positivity-preserving high-order schemes for conservation laws: survey and new developments doi: 10.1098/rspa.2011.0153

The limiter is applied to all scalar variables in their given order using the associated thresholds to determine the minimal acceptable values. The order of the variables is important and might have a strong influence on the robustness.

Trixi.SaveRestartCallback — Type

SaveRestartCallback(; interval=0,
                      save_final_restart=true,
                      output_directory="out")

Save the current numerical solution in a restart file every interval time steps.

Trixi.SaveSolutionCallback — Type

SaveSolutionCallback(; interval=0,
                       save_initial_solution=true,
                       save_final_solution=true,
                       output_directory="out",
                       solution_variables=:primitive)

Save the current numerical solution every interval time steps.

Trixi.SemidiscretizationEulerGravity — Type

SemidiscretizationEulerGravity

A struct containing everything needed to describe a spatial semidiscretization of a the compressible Euler equations with self-gravity, reformulating the Poisson equation for the gravitational potential as steady-state problem of the hyperblic diffusion equations.

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) "A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics" arXiv: 2008.10593

Trixi.SemidiscretizationEulerGravity — Method

SemidiscretizationEulerGravity(semi_euler::SemiEuler, semi_gravity::SemiGravity, parameters)

Construct a semidiscretization of the compressible Euler equations with self-gravity. parameters should be given as ParametersEulerGravity.

Trixi.SemidiscretizationHyperbolic — Type

SemidiscretizationHyperbolic

A struct containing everything needed to describe a spatial semidiscretization of a hyperbolic conservation law.

Trixi.SemidiscretizationHyperbolic — Method

SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver;
                             source_terms=nothing,
                             boundary_conditions=boundary_condition_periodic)

Construct a semidiscretization of a hyperbolic PDE.

Trixi.SteadyStateCallback — Type

SteadyStateCallback(; abstol=1.0e-8, reltol=1.0e-6)

Terminates the integration when the residual_steady_state(du, equations) falls below the threshold specified by abstol, reltol.

Trixi.StepsizeCallback — Type

StepsizeCallback(; cfl=1.0)

Set the time step size according to a CFL condition with CFL number cfl if the time integration method isn't adaptive itself.

Trixi.VolumeIntegralFluxDifferencing — Type

VolumeIntegralFluxDifferencing

Volume integral type for DG methods based on SBP operators and flux differencing using symmetric two-point volume fluxes. Based upon the theory developed by

LeFloch, Mercier, Rohde (2002) Fully Discrete, Entropy Conservative Schemes of Arbitrary Order doi: 10.1137/S003614290240069X

Fisher, Carpenter (2013) High-order entropy stable finite difference schemes for nonlinear conservation laws: Finite domains doi: 10.1016/j.jcp.2013.06.014

Ranocha (2017) Comparison of Some Entropy Conservative Numerical Fluxes for the Euler Equations arXiv: 1701.02264doi: 10.1007/s10915-017-0618-1

Chen, Shu (2017) Entropy stable high order discontinuous Galerkin methods with suitable quadrature rules for hyperbolic conservation laws doi: 10.1016/j.jcp.2017.05.025

Trixi.VolumeIntegralShockCapturingHG — Type

VolumeIntegralShockCapturingHG

Shock-capturing volume integral type for DG methods proposed by

Hennemann, Gassner (2020) "A provably entropy stable subcell shock capturing approach for high order split form DG" arXiv: 2008.12044

Trixi.VolumeIntegralWeakForm — Type

VolumeIntegralWeakForm

The classical weak form volume integral type for DG methods as explained in standard textbooks such as

Kopriva (2009) Implementing Spectral Methods for Partial Differential Equations: Algorithms for Scientists and Engineers doi: 10.1007/978-90-481-2261-5

Base.resize! — Method

resize!(c::AbstractContainer, new_length) -> AbstractContainer

Resize c to contain new_length elements. If new_length is smaller than the current container length, the first new_length elements will be retained. If new_length is larger, the new elements are invalidated.

Trixi.SummaryCallback — Method

SummaryCallback()

Create and return a callback that prints a human-readable summary of the simulation setup at the beginning of a simulation and then resets the timer. When the returned callback is executed directly, the current timer values are shown.

Trixi.TrivialCallback — Method

TrivialCallback()

A callback that does nothing. This can be useful to disable some callbacks easily via trixi_include.

Trixi.boundary_condition_convergence_test — Method

boundary_condition_convergence_test(u_inner, orientation, direction, x, t,
                                    surface_flux_function,
                                    equations::CompressibleEulerEquations1D)

Boundary conditions used for convergence tests in combination with initial_condition_convergence_test and source_terms_convergence_test.

Trixi.boundary_condition_convergence_test — Method

boundary_condition_convergence_test(u_inner, orientation, direction, x, t,
                                    surface_flux_function,
                                    equations::CompressibleEulerEquations2D)

Boundary conditions used for convergence tests in combination with initial_condition_convergence_test and source_terms_convergence_test.

Trixi.boundary_condition_convergence_test — Method

boundary_condition_convergence_test(u_inner, orientation, direction, x, t,
                                    surface_flux_function,
                                    equation::LinearScalarAdvectionEquation1D)

Boundary conditions for initial_condition_convergence_test.

Trixi.boundary_condition_gauss — Method

boundary_condition_gauss(u_inner, orientation, direction, x, t,
                         surface_flux_function,
                         equation::LinearScalarAdvectionEquation1D)

Boundary conditions for initial_condition_gauss.

Trixi.boundary_condition_gauss — Method

boundary_condition_gauss(u_inner, orientation, direction, x, t,
                         surface_flux_function,
                         equation::LinearScalarAdvectionEquation2D)

Boundary conditions for initial_condition_gauss.

Trixi.boundary_condition_linear_x — Method

boundary_condition_linear_x(u_inner, orientation, direction, x, t,
                            surface_flux_function,
                            equation::LinearScalarAdvectionEquation1D)

Boundary conditions for initial_condition_linear_x.

Trixi.boundary_condition_linear_x — Method

boundary_condition_linear_x(u_inner, orientation, direction, x, t,
                            surface_flux_function,
                            equation::LinearScalarAdvectionEquation2D)

Boundary conditions for initial_condition_linear_x.

Trixi.boundary_condition_linear_x_y — Method

boundary_condition_linear_x_y(u_inner, orientation, direction, x, t,
                              surface_flux_function,
                              equation::LinearScalarAdvectionEquation2D)

Boundary conditions for initial_condition_linear_x_y.

Trixi.boundary_condition_linear_y — Method

boundary_condition_linear_y(u_inner, orientation, direction, x, t,
                            surface_flux_function,
                            equation::LinearScalarAdvectionEquation2D)

Boundary conditions for initial_condition_linear_y.

Trixi.boundary_condition_linear_z — Method

boundary_condition_linear_z(u_inner, orientation, direction, x, t,
                            surface_flux_function,
                            equation::LinearScalarAdvectionEquation1D)

Boundary conditions for boundary_condition_linear_z.

Trixi.boundary_condition_sedov_self_gravity — Method

boundary_condition_sedov_self_gravity(u_inner, orientation, direction, x, t,
                                      surface_flux_function,
                                      equations::CompressibleEulerEquations2D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with initial_condition_sedov_self_gravity.

Trixi.boundary_condition_sedov_self_gravity — Method

boundary_condition_sedov_self_gravity(u_inner, orientation, direction, x, t,
                                      surface_flux_function,
                                      equations::CompressibleEulerEquations2D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with initial_condition_sedov_self_gravity.

Trixi.boundary_condition_sedov_self_gravity — Method

boundary_condition_sedov_self_gravity(u_inner, orientation, direction, x, t,
                                      surface_flux_function,
                                      equations::HyperbolicDiffusionEquations2D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with initial_condition_sedov_self_gravity.

Trixi.boundary_condition_sedov_self_gravity — Method

boundary_condition_sedov_self_gravity(u_inner, orientation, direction, x, t,
                                      surface_flux_function,
                                      equations::HyperbolicDiffusionEquations3D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with initial_condition_sedov_self_gravity.

Trixi.calc_error_norms — Method

calc_error_norms([func=(u_node,equations)->u_node,] u_ode, t, analyzer, semi::AbstractSemidiscretization)

Calculate discrete L2 and L∞ error norms of func applied to each nodal variable u_node in u_ode. If no exact solution is available, "errors" are calculated using some reference state and can be useful for regression tests.

Trixi.calc_fast_wavespeed_roe — Method

calc_fast_wavespeed_roe(u_ll, u_rr, direction, equations::IdealGlmMhdEquations2D)

Compute the fast magnetoacoustic wave speed using Roe averages as given by Cargo and Gallice (1997) Roe Matrices for Ideal MHD and Systematic Construction of Roe Matrices for Systems of Conservation Laws DOI: 10.1006/jcph.1997.5773

Trixi.calc_fast_wavespeed_roe — Method

calc_fast_wavespeed_roe(u_ll, u_rr, direction, equations::IdealGlmMhdEquations3D)

Compute the fast magnetoacoustic wave speed using Roe averages as given by

Cargo and Gallice (1997) Roe Matrices for Ideal MHD and Systematic Construction of Roe Matrices for Systems of Conservation Laws DOI: 10.1006/jcph.1997.5773

Trixi.calcflux — Method

calcflux(u, orientation, equations)

Given the conservative variables u, calculate the (physical) flux in spatial direction orientation for the coressponding set of governing equationsorientation is 1, 2, and 3 for the x-, y-, and z-directions, respectively.

Trixi.calcflux_fv! — Method

calcflux_fv!(fstar1_L, fstar1_R, fstar2_L, fstar2_R, u_leftright,
             nonconservative_terms::Val{true}, equations, volume_flux_fv, dg, element)

Calculate the finite volume fluxes inside the elements (with non-conservative terms).

Arguments

fstar1_L::AbstractArray{<:Real, 3}:

fstar1_R::AbstractArray{<:Real, 3}:

fstar2_L::AbstractArray{<:Real, 3}:

fstar2_R::AbstractArray{<:Real, 3}:

u_leftright::AbstractArray{<:Real, 4}

nonconservative_terms::Bool

equations

volume_flux_fv

dg::DGSEM

element::Integer

Trixi.calcflux_fv! — Method

calcflux_fv!(fstar1_L, fstar1_R, fstar2_L, fstar2_R, u_leftright,
             nonconservative_terms::Val{false}, equations, volume_flux_fv, dg, element)

Calculate the finite volume fluxes inside the elements (without non-conservative terms).

Arguments

fstar1_L::AbstractArray{<:Real, 3}:

fstar1_R::AbstractArray{<:Real, 3}:

fstar2_L::AbstractArray{<:Real, 3}:

fstar2_R::AbstractArray{<:Real, 3}:

u_leftright::AbstractArray{<:Real, 4}

nonconservative_terms::Bool

equations

volume_flux_fv

dg::DGSEM

element::Integer

Trixi.compute_coefficients! — Method

compute_coefficients!(u_ode, func, t, semi::AbstractSemidiscretization)

Same as compute_coefficients but stores the result in u_ode.

Trixi.compute_coefficients — Method

compute_coefficients(func, t, semi::AbstractSemidiscretization)

Compute the discrete coefficients of the continuous function func at time t associated with the semidiscretization semi. For example, the discrete coefficients of func for a discontinuous Galerkin spectral element method (DGSEM) are the values of func at the Lobatto-Legendre nodes. Similarly, a classical finite difference method will use the values of func at the nodes of the grid assoociated with the semidiscretization semi.

For semidiscretizations semi associated with an initial condition, func can be omitted to use the given initial condition at time t.

Trixi.convergence_test — Method

convergence_test([mod::Module=Main,] elixir::AbstractString, iterations; kwargs...)

Run iterations Trixi simulations using the setup given in elixir and compute the experimental order of convergence (EOC) in the $L^2$ and $L^\infty$ norm. In each iteration, the initial_refinement_level will be increased by 1. Additional keyword arguments kwargs... and the optional module mod are passed directly to trixi_include.

Trixi.default_example — Method

default_example()

Return the path to an example elixir that can be used to quickly see Trixi in action.

Trixi.examples_dir — Method

examples_dir()

Return the directory where the example files provided with Trixi.jl are located. If Trixi is installed as a regular package (with ]add Trixi), these files are read-only and should not be modified. To find out which files are available, use, e.g., readdir:

Examples

julia> readdir(examples_dir())
5-element Array{String,1}:
 "1d"
 "2d"
 "3d"
 "README.md"
 "paper-self-gravitating-gas-dynamics"

Trixi.flux_central — Method

flux_central(u_ll, u_rr, orientation, equations::AbstractEquations)

The classical central numerical flux f((u_ll) + f(u_rr)) / 2. When this flux is used as volume flux, the discretization is equivalent to the classical weak form DG method (except floating point errors).

Trixi.flux_chandrashekar — Method

flux_chandrashekar(u_ll, u_rr, orientation, equations::CompressibleEulerEquations2D)

Entropy conserving two-point flux by

Chandrashekar (2013) Kinetic Energy Preserving and Entropy Stable Finite Volume Schemes for Compressible Euler and Navier-Stokes Equations DOI: 10.4208/cicp.170712.010313a

Trixi.flux_chandrashekar — Method

flux_chandrashekar(u_ll, u_rr, orientation, equations::CompressibleEulerEquations2D)

Entropy conserving two-point flux by

Chandrashekar (2013) Kinetic Energy Preserving and Entropy Stable Finite Volume Schemes for Compressible Euler and Navier-Stokes Equations DOI: 10.4208/cicp.170712.010313a

Trixi.flux_chandrashekar — Method

flux_chandrashekar(u_ll, u_rr, orientation, equations::CompressibleEulerEquations3D)

Entropy conserving two-point flux by

Chandrashekar (2013) Kinetic Energy Preserving and Entropy Stable Finite Volume Schemes for Compressible Euler and Navier-Stokes Equations DOI: 10.4208/cicp.170712.010313a

Trixi.flux_derigs_etal — Method

flux_derigs_etal(u_ll, u_rr, orientation, equations::IdealGlmMhdEquations2D)

Entropy conserving two-point flux by

Derigs et al. (2018) Ideal GLM-MHD: About the entropy consistent nine-wave magnetic field divergence diminishing ideal magnetohydrodynamics equations DOI: 10.1016/j.jcp.2018.03.002

Trixi.flux_derigs_etal — Method

flux_derigs_etal(u_ll, u_rr, orientation, equations::IdealGlmMhdEquations3D)

Entropy conserving two-point flux by

Derigs et al. (2018) Ideal GLM-MHD: About the entropy consistent nine-wave magnetic field divergence diminishing ideal magnetohydrodynamics equations DOI: 10.1016/j.jcp.2018.03.002

Trixi.flux_hll — Method

flux_hll(u_ll, u_rr, orientation, equations::IdealGlmMhdEquations2D)

HLL flux for ideal GLM-MHD equations like that by Li (2005) An HLLC Riemann solver for magneto-hydrodynamics DOI: 10.1016/j.jcp.2004.08.020

Trixi.flux_hll — Method

flux_hll(u_ll, u_rr, orientation, equations::IdealGlmMhdEquations3D)

HLL flux for ideal GLM-MHD equations like that by

Li (2005) An HLLC Riemann solver for magneto-hydrodynamics DOI: 10.1016/j.jcp.2004.08.020

Trixi.flux_hllc — Method

flux_hllc(u_ll, u_rr, orientation, equations::CompressibleEulerEquations3D)

Computes the HLLC flux (HLL with Contact) for compressible Euler equations developed by E.F. Toro Lecture slides Signal speeds: DOI: 10.1137/S1064827593260140

Trixi.flux_kennedy_gruber — Method

flux_kennedy_gruber(u_ll, u_rr, orientation, equations::CompressibleEulerEquations1D)

Kinetic energy preserving two-point flux by

Kennedy and Gruber (2008) Reduced aliasing formulations of the convective terms within the Navier-Stokes equations for a compressible fluid DOI: 10.1016/j.jcp.2007.09.020

Trixi.flux_kennedy_gruber — Method

flux_kennedy_gruber(u_ll, u_rr, orientation, equations::CompressibleEulerEquations2D)

Kinetic energy preserving two-point flux by

Kennedy and Gruber (2008) Reduced aliasing formulations of the convective terms within the Navier-Stokes equations for a compressible fluid DOI: 10.1016/j.jcp.2007.09.020

Trixi.flux_kennedy_gruber — Method

flux_kennedy_gruber(u_ll, u_rr, orientation, equations::CompressibleEulerEquations3D)

Kinetic energy preserving two-point flux by

Kennedy and Gruber (2008) Reduced aliasing formulations of the convective terms within the Navier-Stokes equations for a compressible fluid DOI: 10.1016/j.jcp.2007.09.020

Trixi.flux_ranocha — Method

flux_ranocha(u_ll, u_rr, orientation, equations::CompressibleEulerEquations2D)

Entropy conserving and kinetic energy preserving two-point flux by

Ranocha (2018) Generalised Summation-by-Parts Operators and Entropy Stability of Numerical Methods for Hyperbolic Balance Laws PhD thesis, TU Braunschweig

See also

Ranocha (2020) Entropy Conserving and Kinetic Energy Preserving Numerical Methods for the Euler Equations Using Summation-by-Parts Operators

Proceedings of ICOSAHOM 2018

Trixi.flux_ranocha — Method

flux_ranocha(u_ll, u_rr, orientation, equations::CompressibleEulerEquations2D)

Entropy conserving and kinetic energy preserving two-point flux by

Ranocha (2018) Generalised Summation-by-Parts Operators and Entropy Stability of Numerical Methods for Hyperbolic Balance Laws PhD thesis, TU Braunschweig

See also

Ranocha (2020) Entropy Conserving and Kinetic Energy Preserving Numerical Methods for the Euler Equations Using Summation-by-Parts Operators Proceedings of ICOSAHOM 2018

Trixi.flux_ranocha — Method

flux_ranocha(u_ll, u_rr, orientation, equations::CompressibleEulerEquations3D)

Entropy conserving and kinetic energy preserving two-point flux by

Ranocha (2018) Generalised Summation-by-Parts Operators and Entropy Stability of Numerical Methods for Hyperbolic Balance Laws PhD thesis, TU Braunschweig

See also

Ranocha (2020) Entropy Conserving and Kinetic Energy Preserving Numerical Methods for the Euler Equations Using Summation-by-Parts Operators Proceedings of ICOSAHOM 2018

Trixi.flux_shima_etal — Method

function flux_shima_etal(u_ll, u_rr, orientation, equations::CompressibleEulerEquations1D)

This flux is is a modification of the original kinetic energy preserving two-point flux by

Kuya, Totani and Kawai (2018) Kinetic energy and entropy preserving schemes for compressible flows by split convective forms DOI: 10.1016/j.jcp.2018.08.058

The modification is in the energy flux to guarantee pressure equilibrium and was developed by

Nao Shima, Yuichi Kuya, Yoshiharu Tamaki, Soshi Kawai (JCP 2020) Preventing spurious pressure oscillations in split convective form discretizations for compressible flows

Trixi.flux_shima_etal — Method

function flux_shima_etal(u_ll, u_rr, orientation, equations::CompressibleEulerEquations2D)

This flux is is a modification of the original kinetic energy preserving two-point flux by

Kuya, Totani and Kawai (2018) Kinetic energy and entropy preserving schemes for compressible flows by split convective forms DOI: 10.1016/j.jcp.2018.08.058

The modification is in the energy flux to guarantee pressure equilibrium and was developed by

Nao Shima, Yuichi Kuya, Yoshiharu Tamaki, Soshi Kawai (JCP 2020) Preventing spurious pressure oscillations in split convective form discretizations for compressible flows

Trixi.flux_shima_etal — Method

function flux_shima_etal(u_ll, u_rr, orientation, equations::CompressibleEulerEquations3D)

This flux is is a modification of the original kinetic energy preserving two-point flux by

Kuya, Totani and Kawai (2018) Kinetic energy and entropy preserving schemes for compressible flows by split convective forms DOI: 10.1016/j.jcp.2018.08.058

The modification is in the energy flux to guarantee pressure equilibrium and was developed by

Nao Shima, Yuichi Kuya, Yoshiharu Tamaki, Soshi Kawai (JCP 2020) Preventing spurious pressure oscillations in split convective form discretizations for compressible flows

Trixi.get_examples — Method

get_examples()

Return a list of all example elixirs that are provided by Trixi.

Trixi.get_name — Method

get_name(x)

Returns a name of x ready for pretty printing. By default, return string(y) if x isa Val{y} and return string(x) otherwise.

Examples

julia> Trixi.get_name("test")
"test"

julia> Trixi.get_name(Val(:test))
"test"

Trixi.init_mpi — Method

init_mpi()

Initialize MPI by calling MPI.Initialized(). The function will check if MPI is already initialized and if yes, do nothing, thus it is safe to call it multiple times.

Trixi.initial_condition_blast_wave — Method

initial_condition_blast_wave(x, t, equations::CompressibleEulerEquations1D)

A medium blast wave taken from

Sebastian Hennemann, Gregor J. Gassner (2020) A provably entropy stable subcell shock capturing approach for high order split form DG arXiv: 2008.12044

Trixi.initial_condition_blast_wave — Method

initial_condition_blast_wave(x, t, equations::CompressibleEulerEquations2D)

A medium blast wave taken from

Sebastian Hennemann, Gregor J. Gassner (2020) A provably entropy stable subcell shock capturing approach for high order split form DG arXiv: 2008.12044

Trixi.initial_condition_blast_wave — Method

initial_condition_blast_wave(x, t, equations::IdealGlmMhdEquations2D)

An MHD blast wave taken from

Dominik Derigs, Gregor J. Gassner, Stefanie Walch & Andrew R. Winters (2018) Entropy Stable Finite Volume Approximations for Ideal Magnetohydrodynamics doi: 10.1365/s13291-018-0178-9

Trixi.initial_condition_blob — Method

initial_condition_blob(x, t, equations::CompressibleEulerEquations2D)

The blob test case taken from

Agertz et al. (2006) Fundamental differences between SPH and grid methods arXiv: astro-ph/0610051

Trixi.initial_condition_blob — Method

initial_condition_blob(x, t, equations::CompressibleEulerEquations3D)

The blob test case taken from

Agertz et al. (2006) Fundamental differences between SPH and grid methods arXiv: astro-ph/0610051

Trixi.initial_condition_constant — Method

initial_condition_constant(x, t, equations::CompressibleEulerEquations1D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_constant — Method

initial_condition_constant(x, t, equations::CompressibleEulerEquations2D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_constant — Method

initial_condition_constant(x, t, equations::CompressibleEulerEquations3D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_constant — Method

initial_condition_constant(x, t, equations::IdealGlmMhdEquations2D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_constant — Method

initialconditionconstant(x, t, equations::IdealGlmMhdEquations3D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_constant — Method

initial_condition_constant(x, t, equations::LinearScalarAdvectionEquation1D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_constant — Method

initial_condition_constant(x, t, equations::LinearScalarAdvectionEquation2D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_constant — Method

initial_condition_constant(x, t, equations::LinearScalarAdvectionEquation1D)

A constant initial condition to test free-stream preservation.

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::CompressibleEulerEquations1D)

A smooth initial condition used for convergence tests in combination with source_terms_convergence_test (and boundary_condition_convergence_test in non-periodic domains).

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::CompressibleEulerEquations2D)

A smooth initial condition used for convergence tests in combination with source_terms_convergence_test (and boundary_condition_convergence_test in non-periodic domains).

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::CompressibleEulerEquations3D)

A smooth initial condition used for convergence tests in combination with source_terms_convergence_test.

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::IdealGlmMhdEquations2D)

An Alfvén wave as smooth initial condition used for convergence tests.

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::IdealGlmMhdEquations3D)

An Alfvén wave as smooth initial condition used for convergence tests.

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::LinearScalarAdvectionEquation1D)

A smooth initial condition used for convergence tests (in combination with boundary_condition_convergence_test in non-periodic domains).

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::LinearScalarAdvectionEquation2D)

A smooth initial condition used for convergence tests.

Trixi.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::LinearScalarAdvectionEquation1D)

A smooth initial condition used for convergence tests.

Trixi.initial_condition_density_pressure_pulse — Method

initial_condition_density_pressure_pulse(x, t, equations::CompressibleEulerEquations2D)

A Gaussian pulse in density and pressure with constant velocity.

Trixi.initial_condition_density_pulse — Method

initial_condition_density_pulse(x, t, equations::CompressibleEulerEquations1D)

A Gaussian pulse in the density with constant velocity and pressure; reduces the compressible Euler equations to the linear advection equations.

Trixi.initial_condition_density_pulse — Method

initial_condition_density_pulse(x, t, equations::CompressibleEulerEquations2D)

A Gaussian pulse in the density with constant velocity and pressure; reduces the compressible Euler equations to the linear advection equations.

Trixi.initial_condition_density_pulse — Method

initial_condition_density_pulse(x, t, equations::CompressibleEulerEquations3D)

A Gaussian pulse in the density with constant velocity and pressure; reduces the compressible Euler equations to the linear advection equations.

Trixi.initial_condition_density_wave — Method

initial_condition_density_wave(x, t, equations::CompressibleEulerEquations1D)

A sine wave in the density with constant velocity and pressure; reduces the compressible Euler equations to the linear advection equations. This setup is the test case for stability of EC fluxes from paper

Gregor J. Gassner, Magnus Svärd, Florian J. Hindenlang (2020) Stability issues of entropy-stable and/or split-form high-order schemes arXiv: 2007.09026

with the following parameters

domain [-1, 1]

mesh = 4x4

polydeg = 5

Trixi.initial_condition_density_wave — Method

initial_condition_density_wave(x, t, equations::CompressibleEulerEquations2D)

A sine wave in the density with constant velocity and pressure; reduces the compressible Euler equations to the linear advection equations. This setup is the test case for stability of EC fluxes from paper

Gregor J. Gassner, Magnus Svärd, Florian J. Hindenlang (2020) Stability issues of entropy-stable and/or split-form high-order schemes arXiv: 2007.09026

with the following parameters

domain [-1, 1]

mesh = 4x4

polydeg = 5

Trixi.initial_condition_eoc_test_coupled_euler_gravity — Method

initial_condition_eoc_test_coupled_euler_gravity(x, t, equations::CompressibleEulerEquations2D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with source_terms_eoc_test_coupled_euler_gravity or source_terms_eoc_test_euler.

Trixi.initial_condition_eoc_test_coupled_euler_gravity — Method

initial_condition_eoc_test_coupled_euler_gravity(x, t, equations::CompressibleEulerEquations3D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with source_terms_eoc_test_coupled_euler_gravity or source_terms_eoc_test_euler.

Trixi.initial_condition_eoc_test_coupled_euler_gravity — Method

initial_condition_eoc_test_coupled_euler_gravity(x, t, equations::HyperbolicDiffusionEquations2D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with source_terms_harmonic.

Trixi.initial_condition_eoc_test_coupled_euler_gravity — Method

initial_condition_eoc_test_coupled_euler_gravity(x, t, equations::HyperbolicDiffusionEquations3D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with source_terms_harmonic.

Trixi.initial_condition_gauss — Method

initial_condition_gauss(x, t, equations::LinearScalarAdvectionEquation1D)

A Gaussian pulse used together with boundary_condition_gauss.

Trixi.initial_condition_gauss — Method

initial_condition_gauss(x, t, equation::LinearScalarAdvectionEquation2D)

A Gaussian pulse used together with boundary_condition_gauss.

Trixi.initial_condition_gauss — Method

initial_condition_gauss(x, t, equations::LinearScalarAdvectionEquation1D)

A Gaussian pulse.

Trixi.initial_condition_isentropic_vortex — Method

initial_condition_isentropic_vortex(x, t, equations::CompressibleEulerEquations2D)

The classical isentropic vortex test case of

Chi-Wang Shu (1997) Essentially Non-Oscillatory and Weighted Essentially Non-Oscillatory Schemes for Hyperbolic Conservation Laws NASA/CR-97-206253

Trixi.initial_condition_khi — Method

initial_condition_khi(x, t, equations::CompressibleEulerEquations2D)

The classical Kelvin-Helmholtz instability based on

https://rsaa.anu.edu.au/research/established-projects/fyris/2-d-kelvin-helmholtz-test

Trixi.initial_condition_linear_x — Method

initial_condition_linear_x(x, t, equations::LinearScalarAdvectionEquation1D)

A linear function of x[1] used together with boundary_condition_linear_x.

Trixi.initial_condition_linear_x — Method

initial_condition_linear_x(x, t, equations::LinearScalarAdvectionEquation2D)

A linear function of x[1] used together with boundary_condition_linear_x.

Trixi.initial_condition_linear_x_y — Method

initial_condition_linear_x_y(x, t, equations::LinearScalarAdvectionEquation2D)

A linear function of x[1] + x[2] used together with boundary_condition_linear_x_y.

Trixi.initial_condition_linear_y — Method

initial_condition_linear_y(x, t, equations::LinearScalarAdvectionEquation2D)

A linear function of x[1] used together with boundary_condition_linear_y.

Trixi.initial_condition_linear_z — Method

initial_condition_linear_z(x, t, equations::LinearScalarAdvectionEquation1D)

A linear function of x[3] used together with boundary_condition_linear_z.

Trixi.initial_condition_medium_sedov_blast_wave — Method

initial_condition_medium_sedov_blast_wave(x, t, equations::CompressibleEulerEquations2D)

The Sedov blast wave setup based on Flash

http://flash.uchicago.edu/site/flashcode/usersupport/flashug_devel/node184.html#SECTION010114000000000000000

with smaller strength of the initial discontinuity.

Trixi.initial_condition_orszag_tang — Method

initial_condition_orszag_tang(x, t, equations::IdealGlmMhdEquations2D)

The classical Orszag-Tang vortex test case. Here, the setup is taken from

Dominik Derigs, Gregor J. Gassner, Stefanie Walch & Andrew R. Winters (2018) Entropy Stable Finite Volume Approximations for Ideal Magnetohydrodynamics doi: 10.1365/s13291-018-0178-9

Trixi.initial_condition_orszag_tang — Method

initial_condition_orszag_tang(x, t, equations::IdealGlmMhdEquations3D)

The classical Orszag-Tang vortex test case. Here, the setup is taken from

Dominik Derigs, Gregor J. Gassner, Stefanie Walch & Andrew R. Winters (2018) Entropy Stable Finite Volume Approximations for Ideal Magnetohydrodynamics doi: 10.1365/s13291-018-0178-9

Trixi.initial_condition_pressure_pulse — Method

initial_condition_pressure_pulse(x, t, equations::CompressibleEulerEquations2D)

A Gaussian pulse in the pressure with constant velocity and density.

Trixi.initial_condition_rotor — Method

initial_condition_rotor(x, t, equations::IdealGlmMhdEquations2D)

The classical MHD rotor test case. Here, the setup is taken from

Dominik Derigs, Gregor J. Gassner, Stefanie Walch & Andrew R. Winters (2018) Entropy Stable Finite Volume Approximations for Ideal Magnetohydrodynamics doi: 10.1365/s13291-018-0178-9

Trixi.initial_condition_sedov_blast_wave — Method

initial_condition_sedov_blast_wave(x, t, equations::CompressibleEulerEquations1D)

The Sedov blast wave setup based on Flash

http://flash.uchicago.edu/site/flashcode/usersupport/flashug_devel/node184.html#SECTION010114000000000000000

Trixi.initial_condition_sedov_blast_wave — Method

initial_condition_sedov_blast_wave(x, t, equations::CompressibleEulerEquations2D)

The Sedov blast wave setup based on Flash

http://flash.uchicago.edu/site/flashcode/usersupport/flashug_devel/node184.html#SECTION010114000000000000000

Trixi.initial_condition_sedov_blast_wave — Method

initial_condition_sedov_blast_wave(x, t, equations::CompressibleEulerEquations3D)

The Sedov blast wave setup based on Flash

http://flash.uchicago.edu/site/flashcode/usersupport/flashug_devel/node184.html#SECTION010114000000000000000

Trixi.initial_condition_sedov_self_gravity — Method

initial_condition_sedov_self_gravity(x, t, equations::CompressibleEulerEquations2D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with boundary_condition_sedov_self_gravity.

Trixi.initial_condition_sedov_self_gravity — Method

initial_condition_sedov_self_gravity(x, t, equations::CompressibleEulerEquations3D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with boundary_condition_sedov_self_gravity.

Trixi.initial_condition_sedov_self_gravity — Method

initial_condition_sedov_self_gravity(x, t, equations::HyperbolicDiffusionEquations2D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with boundary_condition_sedov_self_gravity.

Trixi.initial_condition_sedov_self_gravity — Method

initial_condition_sedov_self_gravity(x, t, equations::HyperbolicDiffusionEquations3D)

Adaptation of the Sedov blast wave with self-gravity taken from

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

based on

http://flash.uchicago.edu/site/flashcode/usersupport/flash4ug_4p62/node184.html#SECTION010114000000000000000

Should be used together with boundary_condition_sedov_self_gravity.

Trixi.initial_condition_sin — Method

initial_condition_sin(x, t, equations::LinearScalarAdvectionEquation1D)

A sine wave in the conserved variable.

Trixi.initial_condition_sin — Method

initial_condition_sin(x, t, equations::LinearScalarAdvectionEquation1D)

A sine wave in the conserved variable.

Trixi.initial_condition_sin_sin — Method

initial_condition_sin_sin(x, t, equations::LinearScalarAdvectionEquation2D)

A sine wave in the conserved variable.

Trixi.initial_condition_taylor_green_vortex — Method

initial_condition_taylor_green_vortex(x, t, equations::CompressibleEulerEquations3D)

The classical inviscid Taylor-Green vortex.

Trixi.initial_condition_weak_blast_wave — Method

initial_condition_weak_blast_wave(x, t, equations::CompressibleEulerEquations1D)

A weak blast wave taken from

Sebastian Hennemann, Gregor J. Gassner (2020) A provably entropy stable subcell shock capturing approach for high order split form DG arXiv: 2008.12044

Trixi.initial_condition_weak_blast_wave — Method

initial_condition_weak_blast_wave(x, t, equations::CompressibleEulerEquations2D)

A weak blast wave taken from

Sebastian Hennemann, Gregor J. Gassner (2020) A provably entropy stable subcell shock capturing approach for high order split form DG arXiv: 2008.12044

Trixi.initial_condition_weak_blast_wave — Method

initial_condition_weak_blast_wave(x, t, equations::CompressibleEulerEquations3D)

A weak blast wave taken from

Sebastian Hennemann, Gregor J. Gassner (2020) A provably entropy stable subcell shock capturing approach for high order split form DG arXiv: 2008.12044

Trixi.initial_condition_weak_blast_wave — Method

initial_condition_weak_blast_wave(x, t, equations::IdealGlmMhdEquations2D)

A weak blast wave adapted from

Sebastian Hennemann, Gregor J. Gassner (2020) A provably entropy stable subcell shock capturing approach for high order split form DG arXiv: 2008.12044

Trixi.initial_condition_weak_blast_wave — Method

initial_condition_weak_blast_wave(x, t, equations::IdealGlmMhdEquations3D)

A weak blast wave adapted from

Sebastian Hennemann, Gregor J. Gassner (2020) A provably entropy stable subcell shock capturing approach for high order split form DG arXiv: 2008.12044

Trixi.integrate — Method

integrate([func=(u_node,equations)->u_node,] u_ode::AbstractVector, semi::AbstractSemidiscretization; normalize=true)

Call func(u_node, equations) for each vector of nodal variables u_node in u_ode and integrate the result using a quadrature associated with the semidiscretization semi.

If normalize is true, the result is divided by the total volume of the computational domain.

Trixi.integrate_via_indices — Method

integrate_via_indices(func, u_ode::AbstractVector, semi::AbstractSemidiscretization, args...; normalize=true)

Call func(u, i..., element, equations, solver, args...) for all nodal indices i..., element and integrate the result using a quadrature associated with the semidiscretization semi.

If normalize is true, the result is divided by the total volume of the computational domain.

Trixi.jacobian_fd — Method

jacobian_fd(semi::AbstractSemidiscretization;
            t0=zero(real(semi)),
            u0_ode=compute_coefficients(t0, semi))

Uses the right-hand side operator of the semidiscretization semi and simple second order finite difference to compute the Jacobian J of the operator. The linearization state is u0_ode at time t0.

Trixi.linear_structure — Method

linear_structure(semi::AbstractSemidiscretization;
                 t0=zero(real(semi)))

Wraps the right-hand side operator of the semidiscretization semi at time t0 as an affine-linear operator given by a linear operator A and a vector b.

Trixi.load_mesh — Method

load_mesh(restart_file::AbstractString; n_cells_max)

Load the mesh from the restart_file.

Trixi.load_time — Method

load_time(restart_file::AbstractString)

Load the time saved in a restart_file.

Trixi.multiply_dimensionwise — Method

multiply_dimensionwise(matrix::AbstractMatrix, data_in::AbstractArray{<:Any, NDIMS+1})

Multiply the array data_in by matrix in each coordinate direction, where data_in is assumed to have the first coordinate for the number of variables and the remaining coordinates are multiplied by matrix.

Trixi.ndofs — Method

ndofs(semi::AbstractSemidiscretization)

Return the number of degrees of freedom associated with each scalar variable.

Trixi.noncons_interface_flux — Method

noncons_interface_flux(u_left, u_right, orientation, mode, equations::IdealGlmMhdEquations2D)

Strong form of non-conservative flux on a surface (Powell and GLM terms)

\[phi^L 1/2 (B^L + B^R)_{normal} - phi^L B^L+{normal} = phi^L 1/2 (B^R - B^L)_{normal}\]

Note

The non-conservative interface flux depends on the discretization. Following "modes" are available:

:weak: 'weak' formulation of split DG already includes the contribution $-1/2 (phi^L B^L_{normal})$ so this mode only adds $1/2 (phi^L B^R_{normal})$, analogously for the Galilean nonconservative term

:whole: This mode adds the whole non-conservative term: phi^L 1/2 (B^R-B^L)

:inner: This mode adds the split-form DG volume integral contribution: This is equivalent to $(2)-(1) - 1/2 (phi^L B^L)$

Warning

This is non-unique along an interface! The normal direction is super important.

Trixi.pure_and_blended_element_ids! — Method

pure_and_blended_element_ids!(element_ids_dg, element_ids_dgfv, alpha, dg, cache)

Given blending factors alpha and the solver dg, fill element_ids_dg with the IDs of elements using a pure DG scheme and element_ids_dgfv with the IDs of elements using a blended DG-FV scheme.

Trixi.residual_steady_state — Method

residual_steady_state(du, ::AbstractHyperbolicDiffusionEquations)

Used to determine the termination criterion of a SteadyStateCallback. For hyperbolic diffusion, this checks convergence of the potential $\phi$.

Trixi.semidiscretize — Method

semidiscretize(semi::AbstractSemidiscretization, tspan, restart_file::AbstractString)

Wrap the semidiscretization semi as an ODE problem in the time interval tspan that can be passed to solve from the SciML ecosystem. The initial condition etc. is taken from the restart_file.

Trixi.semidiscretize — Method

semidiscretize(semi::AbstractSemidiscretization, tspan)

Wrap the semidiscretization semi as an ODE problem in the time interval tspan that can be passed to solve from the SciML ecosystem.

Trixi.source_terms_convergence_test — Method

source_terms_convergence_test(u, x, t, equations::CompressibleEulerEquations1D)

Source terms used for convergence tests in combination with initial_condition_convergence_test (and boundary_condition_convergence_test in non-periodic domains).

Trixi.source_terms_convergence_test — Method

source_terms_convergence_test(u, x, t, equations::CompressibleEulerEquations2D)

Source terms used for convergence tests in combination with initial_condition_convergence_test (and boundary_condition_convergence_test in non-periodic domains).

Trixi.source_terms_convergence_test — Method

source_terms_convergence_test(u, x, t, equations::CompressibleEulerEquations3D)

Source terms used for convergence tests in combination with initial_condition_convergence_test.

Trixi.source_terms_eoc_test_coupled_euler_gravity — Method

source_terms_eoc_test_coupled_euler_gravity(u, x, t, equations::CompressibleEulerEquations2D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with initial_condition_eoc_test_coupled_euler_gravity.

Trixi.source_terms_eoc_test_coupled_euler_gravity — Method

source_terms_eoc_test_coupled_euler_gravity(u, x, t, equations::CompressibleEulerEquations3D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with initial_condition_eoc_test_coupled_euler_gravity.

Trixi.source_terms_eoc_test_euler — Method

source_terms_eoc_test_euler(u, x, t, equations::CompressibleEulerEquations2D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with initial_condition_eoc_test_coupled_euler_gravity.

Trixi.source_terms_eoc_test_euler — Method

source_terms_eoc_test_euler(u, x, t, equations::CompressibleEulerEquations3D)

Setup used for convergence tests of the Euler equations with self-gravity used in

Michael Schlottke-Lakemper, Andrew R. Winters, Hendrik Ranocha, Gregor J. Gassner (2020) A purely hyperbolic discontinuous Galerkin approach for self-gravitating gas dynamics arXiv: 2008.10593

in combination with initial_condition_eoc_test_coupled_euler_gravity.

Note

This method is to be used for testing pure Euler simulations with analytic self-gravity. If you intend to do coupled Euler-gravity simulations, you need to use source_terms_eoc_test_coupled_euler_gravity instead.

Trixi.trixi_include — Method

trixi_include([mod::Module=Main,] elixir::AbstractString; kwargs...)

include the file elixir and evaluate its content in the global scope of module mod. You can override specific assignments in elixir by supplying keyword arguments. It's basic purpose is to make it easier to modify some parameters while running Trixi from the REPL. Additionally, this is used in tests to reduce the computational burden for CI while still providing examples with sensible default values for users.

Examples

julia> trixi_include(@__MODULE__, default_example(), tspan=(0.0, 0.1))
[...]

julia> sol.t[end]
0.1