Trixi API

Trixi.Trixi — Module

Trixi

Trixi.jl is a flexible numerical simulation framework for partial differential equations. It is based on a two-dimensional hierarchical mesh (quadtree) and supports several governing equations such as the compressible Euler equations, magnetohydrodynamics equations, or hyperbolic diffusion equations. Trixi is written in Julia and aims to be easy to use and extend also for new or inexperienced users.

See also: trixi-framework/Trixi.jl

Trixi.ndim — Constant

ndim

Specify the number of spatial dimensions.

Always use ndim instead of hard-coding the literal 2 when referring to the number of dimensions. This makes code easier to understand, since it adds a meaning to the number.

Trixi.CompressibleEulerEquations — Type

CompressibleEulerEquations

The compressible Euler equations for an ideal gas in two space dimensions.

Trixi.HyperbolicDiffusionEquations — Type

HyperbolicDiffusionEquations

The linear hyperbolic diffusion equations in two space dimensions. A description of this system can be found in Sec. 2.5 of the book "I Do Like CFD, Too: Vol 1". The book is freely available at http://www.cfdbooks.com/ and further analysis can be found in the paper by Nishikawa DOI: 10.1016/j.jcp.2007.07.029

Trixi.IdealGlmMhdEquations — Type

IdealGlmMhdEquations

The ideal compressible GLM-MHD equations in two space dimensions.

Trixi.LinearScalarAdvectionEquation — Type

LinearScalarAdvectionEquation

The linear scalar advection equation

\[\partial_t u + a_1 \partial_1 u + a_2 \partial_2 u = 0\]

in two space dimensions with constant velocity a.

Base.resize! — Method

resize!(c::AbstractContainer, new_length) -> AbstractContainer

Resize c to contain new_length elements. If new_length is smaller than the current container length, the first new_length elements will be retained. If new_length is larger, the new elements are invalidated.

Trixi.analyze_solution — Method

analyze_solution(dg::Dg, mesh::TreeMesh, time, dt, step, runtime_absolute, runtime_relative)

Calculate error norms and other analysis quantities to analyze the solution during a simulation, and return the L2 and Linf errors. dg and mesh are the DG and the mesh instance, respectively. time, dt, and step refer to the current simulation time, the last time step size, and the current time step count. The run time (in seconds) is given in runtime_absolute, while the performance index is specified in runtime_relative.

Note: Keep order of analysis quantities in sync with save_analysis_header when adding or changing quantities.

Trixi.calcflux_fv! — Method

calcflux_fv!(fstar1, fstar2, u_leftright, u, element_id, dg::Dg)

Calculate the finite volume fluxes inside the elements.

Arguments

fstar1::AbstractArray{T} where T<:Real:

fstar2::AbstractArray{T} where T<:Real

dg::Dg

u::AbstractArray{T} where T<:Real

element_id::Integer

Trixi.convtest — Method

convtest(parameters_file, iterations; parameters...)

Run multiple Trixi simulations with the parameters in parameters_file and compute the experimental order of convergence (EOC) in the $L^2$ and $L^\infty$ norm. The number of runs is specified by iterations and in each run the initial refinement level will be increased by 1. Parameters can be overriden by specifying them as additional keyword arguments, which are passed to the respective call to run..

Trixi.flux_central — Method

flux_central(u_ll, u_rr, orientation, equation::AbstractEquation)

The classical central numerical flux f((u_ll) + f(u_rr)) / 2. When this flux is used as volume flux, the discretization is equivalent to the classical weak form DG method (except floating point errors).

Trixi.flux_chandrashekar — Method

flux_chandrashekar(u_ll, u_rr, orientation, equation::CompressibleEulerEquations)

Entropy conserving two-point flux by Chandrashekar (2013) Kinetic Energy Preserving and Entropy Stable Finite Volume Schemes for Compressible Euler and Navier-Stokes Equations DOI: 10.4208/cicp.170712.010313a

Trixi.flux_derigs_etal — Method

flux_derigs_etal(u_ll, u_rr, orientation, equation::IdealGlmMhdEquations)

Entropy conserving two-point flux by Derigs et al. (2018) Ideal GLM-MHD: About the entropy consistent nine-wave magnetic field divergence diminishing ideal magnetohydrodynamics equations DOI: 10.1016/j.jcp.2018.03.002

Trixi.flux_kennedy_gruber — Method

flux_kennedy_gruber(u_ll, u_rr, orientation, equation::CompressibleEulerEquations)

Kinetic energy preserving two-point flux by Kennedy and Gruber (2008) Reduced aliasing formulations of the convective terms within the Navier-Stokes equations for a compressible fluid DOI: 10.1016/j.jcp.2007.09.020

Trixi.flux_kuya_etal — Method

function flux_kuya_etal(u_ll, u_rr, orientation, equation::CompressibleEulerEquations)

Kinetic energy preserving two-point flux with pressure oscillation fix by Kuya, Totani and Kawai (2018) Kinetic energy and entropy preserving schemes for compressible flows by split convective forms DOI: 10.1016/j.jcp.2018.08.058

Trixi.flux_ranocha — Method

flux_ranocha(u_ll, u_rr, orientation, equation::CompressibleEulerEquations)

Entropy conserving and kinetic energy preserving two-point flux by Ranocha (2018) Generalised Summation-by-Parts Operators and Entropy Stability of Numerical Methods for Hyperbolic Balance Laws PhD thesis, TU Braunschweig

Trixi.get_name — Method

get_name(x)

Returns a name of x ready for pretty printing. By default, return string(y) if x isa Val{y} and return string(x) otherwise.

Examples

julia> Trixi.get_name("test")
"test"

julia> Trixi.get_name(Val(:test))
"test"

Trixi.integrate — Method

integrate(func, u, dg::Dg; normalize=true)
integrate(u, dg::Dg; normalize=true)

Call function func for each DG node and integrate the result over the computational domain.

The function func is called as func(u_local) for each volume node (i, j) and each element_id, where u_local is an SVectorized copy of u[:, i, j, element_id]. If normalize is true, the result is divided by the total volume of the computational domain. If func is omitted, it defaults to identity.

Examples

Calculate the integral over all conservative variables:

state_integrals = integrate(dg.elements.u, dg)

Trixi.integrate — Method

integrate(func, dg::Dg, args...; normalize=true)

Call function func for each DG node and integrate the result over the computational domain.

The function func is called as func(i, j, element_id, dg, args...) for each volume node (i, j) and each element_id. Additional positional arguments args... are passed along as well. If normalize is true, the result is divided by the total volume of the computational domain.

Examples

Calculate the integral of the time derivative of the entropy, i.e., ∫(∂S/∂t)dΩ = ∫(∂S/∂u ⋅ ∂u/∂t)dΩ:

# Compute entropy variables
entropy_vars = cons2entropy(...)

# Calculate integral of entropy time derivative
dsdu_ut = integrate(dg, entropy_vars, dg.elements.u_t) do i, j, element_id, dg, entropy_vars, u_t
  sum(entropy_vars[:, i, j, element_id] .* u_t[:, i, j, element_id])
end

Trixi.pure_and_blended_element_ids! — Method

pure_and_blended_element_ids!(element_ids_dg, element_ids_dgfv, alpha, dg)

Given blending factors alpha and the solver dg, fill element_ids_dg with the IDs of elements using a pure DG scheme and element_ids_dgfv with the IDs of elements using a blended DG-FV scheme.

Trixi.riemann! — Method

riemann!(destination, u_interfaces_left, u_interfaces_right, interface_id, orientations, dg::Dg)

Calculate the surface flux across interface with different states given by u_interfaces_left, u_interfaces_right on both sides (EC mortar version).

Arguments

destination::AbstractArray{T,3} where T<:Real: The array of surface flux values (updated inplace).

u_interfaces_left::AbstractArray{T,3} where T<:Real`

u_interfaces_right::AbstractArray{T,3} where T<:Real`

interface_id::Integer

orientations::Vector{T} where T<:Integer

dg::Dg

Trixi.riemann! — Method

riemann!(destination, u_interfaces, interface_id, orientations, dg::Dg)

Calculate the surface flux across interface with different states given by u_interfaces_left, u_interfaces_right on both sides (interface version).

Arguments

destination::AbstractArray{T,2} where T<:Real: The array of surface flux values (updated inplace).

u_interfaces::AbstractArray{T,4} where T<:Real`

interface_id::Integer

orientations::Vector{T} where T<:Integer

dg::Dg

Trixi.run — Function

run(parameters_file=nothing; verbose=false, args=nothing, refinement_level_increment=0, parameters...)

Run a Trixi simulation with the parameters in parameters_file. Parameters can be overriden by specifying them as keyword arguments (see examples).

If verbose is true, additional output will be generated on the terminal that may help with debugging. If args is given, it should be an ARGS-like array of strings that holds command line arguments, and will be interpreted by the parse_commandline_arguments function. In this case, the values of parameters_file and verbose are ignored. If a value for refinement_level_increment is given, initial_refinement_level will be increased by this value before running the simulation (mostly used by EOC analysis).

Examples

julia> Trixi.run("examples/parameters.toml", verbose=true)
[...]

Without changing the parameters file we can start a simulation with N = 1 and t_end = 0.5 as follows:

julia> Trixi.run("examples/parameters.toml", N=1, t_end=0.5)
[...]

Trixi.save_analysis_header — Method

save_analysis_header(filename, quantities, equation)

Truncate file filename and save a header with the names of the quantities quantities that will subsequently written to filename by analyze_solution. Since some quantities are equation-specific, the system of equations instance is passed in equation.

Note: Keep order of analysis quantities in sync with analyze_solution when adding or changing quantities.

Trixi.setparameter — Method

setparameter(name::String, value)

Set parameter with the specified name to the specified value.

Trixi.timestep_2N! — Method

timestep_2N!(solver, t, dt, a, b, c)

Perform one timestep using an explicit Runge-Kutta method of the low-storage class 2N of Williamson.

Trixi.timestep_3Sstar! — Method

timestep_3Sstar!(solver, t, dt, gamma1, gamma2, gamma3, beta, delta, c)

Perform one timestep using an explicit Runge-Kutta method of the low-storage class 3Sstar of Ketcheson.

Trixi.timestep_carpenter_kennedy_erk43_2N! — Method

timestep_carpenter_kennedy_erk43_2N!(solver::AbstractSolver, t, dt)

Carpenter, Kennedy (1994) Third order 2N storage RK schemes with error control

Trixi.timestep_carpenter_kennedy_erk54_2N! — Method

timestep_carpenter_kennedy_erk54_2N!(solver::AbstractSolver, t, dt)

Carpenter, Kennedy (1994) Fourth order 2N storage RK schemes, Solution 3

Trixi.timestep_hyp_diff_N3_erk52_3Sstar! — Method

timestep_hyp_diff_N3_erk52_3Sstar!(solver::AbstractSolver, t, dt)

Five stage, second-order acurate explicit Runge-Kutta scheme with stability region optimized for the hyperbolic diffusion equation with LLF flux and polynomials of degree N=3.

Trixi.timestep_parsani_ketcheson_deconinck_erk32_3Sstar! — Method

timestep_parsani_ketcheson_deconinck_erk32_3Sstar!(solver::AbstractSolver, t, dt)

Parsani, Ketcheson, Deconinck (2013) Optimized explicit RK schemes for the spectral difference method applied to wave propagation problems DOI: 10.1137/120885899

Trixi.timestep_parsani_ketcheson_deconinck_erk94_3Sstar! — Method

timestep_parsani_ketcheson_deconinck_erk94_3Sstar!(solver::AbstractSolver, t, dt)

Parsani, Ketcheson, Deconinck (2013) Optimized explicit RK schemes for the spectral difference method applied to wave propagation problems DOI: 10.1137/120885899