Trixi2Img API

Trixi2Img.run — Method

run(; args=nothing, kwargs...)

Convert Trixi-generated output files to image files (PNG or PDF).

If args is given, it should be an ARGS-like array of strings that holds command line arguments, and will be interpreted by the ArgParse module. If args is omitted, you can supply all command line arguments via keyword arguments. In this case, you have to provide at least one input file path in the datafile variable.

Examples

julia> Trixi2Img.run(datafile="out/solution_000000.h5")
[...]