Trixi2Vtk API

Trixi2Vtk.run — Method

run(; args=nothing, kwargs...)

Convert Trixi-generated output files to VTK files (VTU or VTI).

If args is given, it should be an ARGS-like array of strings that holds command line arguments, and will be interpreted by the ArgParse module. If args is omitted, you can supply all command line arguments via keyword arguments. In this case, you have to provide at least one input file path in the filename variable.

Examples

julia> Trixi2Vtk.run(filename="out/solution_000000.h5")
[...]