Trixi.jl API

Trixi.Trixi — Module

Trixi

Trixi.jl is a numerical simulation framework for hyperbolic conservation laws. A key objective for the framework is to be useful to both scientists and students. Therefore, next to having an extensible design with a fast implementation, Trixi is focused on being easy to use for new or inexperienced users, including the installation and postprocessing procedures.

See also: trixi-framework/Trixi.jl

Trixi.CompressibleEulerEquations2D — Type

CompressibleEulerEquations2D

The compressible Euler equations for an ideal gas in two space dimensions.

Trixi.CompressibleEulerEquations3D — Type

CompressibleEulerEquations3D

The compressible Euler equations for an ideal gas in three space dimensions.

Trixi.HyperbolicDiffusionEquations2D — Type

HyperbolicDiffusionEquations2D

The linear hyperbolic diffusion equations in two space dimensions. A description of this system can be found in Sec. 2.5 of the book "I Do Like CFD, Too: Vol 1". The book is freely available at http://www.cfdbooks.com/ and further analysis can be found in the paper by Nishikawa DOI: 10.1016/j.jcp.2007.07.029

Trixi.HyperbolicDiffusionEquations3D — Type

HyperbolicDiffusionEquations3D

The linear hyperbolic diffusion equations in three space dimensions. A description of this system can be found in Sec. 2.5 of the book "I Do Like CFD, Too: Vol 1". The book is freely available at http://www.cfdbooks.com/ and further analysis can be found in the paper by Nishikawa DOI: 10.1016/j.jcp.2007.07.029

Trixi.IdealGlmMhdEquations2D — Type

IdealGlmMhdEquations2D

The ideal compressible GLM-MHD equations in two space dimensions.

Trixi.IdealGlmMhdEquations3D — Type

IdealGlmMhdEquations3D

The ideal compressible GLM-MHD equations in three space dimensions.

Trixi.LinearScalarAdvectionEquation2D — Type

LinearScalarAdvectionEquation2D

The linear scalar advection equation

\[\partial_t u + a_1 \partial_1 u + a_2 \partial_2 u = 0\]

in two space dimensions with constant velocity a.

Trixi.LinearScalarAdvectionEquation3D — Type

LinearScalarAdvectionEquation3D

The linear scalar advection equation

\[\partial_t u + a_1 \partial_1 u + a_2 \partial_2 u + a_3 \partial_3 u = 0\]

in two space dimensions with constant velocity a.

Base.resize! — Method

resize!(c::AbstractContainer, new_length) -> AbstractContainer

Resize c to contain new_length elements. If new_length is smaller than the current container length, the first new_length elements will be retained. If new_length is larger, the new elements are invalidated.

Trixi.analyze_solution — Method

analyze_solution(dg::Dg2D, mesh::TreeMesh, time, dt, step, runtime_absolute, runtime_relative)

Calculate error norms and other analysis quantities to analyze the solution during a simulation, and return the L2 and Linf errors. dg and mesh are the DG and the mesh instance, respectively. time, dt, and step refer to the current simulation time, the last time step size, and the current time step count. The run time (in seconds) is given in runtime_absolute, while the performance index is specified in runtime_relative.

Note: Keep order of analysis quantities in sync with save_analysis_header when adding or changing quantities.

Trixi.analyze_solution — Method

analyze_solution(dg::Dg3D, mesh::TreeMesh, time, dt, step, runtime_absolute, runtime_relative)

Calculate error norms and other analysis quantities to analyze the solution during a simulation, and return the L2 and Linf errors. dg and mesh are the DG and the mesh instance, respectively. time, dt, and step refer to the current simulation time, the last time step size, and the current time step count. The run time (in seconds) is given in runtime_absolute, while the performance index is specified in runtime_relative.

Note: Keep order of analysis quantities in sync with save_analysis_header when adding or changing quantities.

Trixi.calc_error_norms — Function

calc_error_norms([func=(u,equation)->u,] solver, t)

Calculate the discrete L2 and L∞ errors of func applied to the conservative variables of the problem encapsulated by solver at time t, where func is called as func(u, equation).

Trixi.calc_fast_wavespeed_roe — Method

calc_fast_wavespeed_roe(u_ll, u_rr, direction, equation::IdealGlmMhdEquations2D)

Compute the fast magnetoacoustic wave speed using Roe averages as given by Cargo and Gallice (1997) Roe Matrices for Ideal MHD and Systematic Construction of Roe Matrices for Systems of Conservation Laws DOI: 10.1006/jcph.1997.5773

Trixi.calc_fast_wavespeed_roe — Method

calc_fast_wavespeed_roe(u_ll, u_rr, direction, equation::IdealGlmMhdEquations3D)

Compute the fast magnetoacoustic wave speed using Roe averages as given by Cargo and Gallice (1997) Roe Matrices for Ideal MHD and Systematic Construction of Roe Matrices for Systems of Conservation Laws DOI: 10.1006/jcph.1997.5773

Trixi.calc_fstar! — Method

calc_fstar!(destination, u_interfaces_left, u_interfaces_right, interface_id, orientations, dg::Dg2D)

Calculate the surface flux across interface with different states given by u_interfaces_left, u_interfaces_right on both sides (EC mortar version).

Arguments

destination::AbstractArray{T,3} where T<:Real: The array of surface flux values (updated inplace).

u_interfaces_left::AbstractArray{T,3} where T<:Real`

u_interfaces_right::AbstractArray{T,3} where T<:Real`

interface_id::Integer

orientations::Vector{T} where T<:Integer

dg::Dg2D

Trixi.calc_fstar! — Method

calc_fstar!(destination, u_interfaces, interface_id, orientations, dg::Dg2D)

Calculate the surface flux across interface with different states given by u_interfaces_left, u_interfaces_right on both sides (interface version).

Arguments

destination::AbstractArray{T,2} where T<:Real: The array of surface flux values (updated inplace).

u_interfaces::AbstractArray{T,4} where T<:Real`

interface_id::Integer

orientations::Vector{T} where T<:Integer

dg::Dg2D

Trixi.calc_fstar! — Method

calc_fstar!(destination, u_interfaces, interface_id, orientations, dg::Dg3D)

Calculate the surface flux across interface with different states given by u_interfaces_left, u_interfaces_right on both sides. Only used by calc_interface_flux! for interfaces with non-conservative terms and for L2 mortars.

Arguments

destination::AbstractArray{T,3} where T<:Real: The array of surface flux values (updated inplace).

u_interfaces::AbstractArray{T,5} where T<:Real`

interface_id::Integer

orientations::Vector{T} where T<:Integer

dg::Dg3D

Trixi.calcflux_fv! — Method

calcflux_fv!(fstar1, fstar2, fstar3, u_leftright, u, element_id, dg::Dg3D)

Calculate the finite volume fluxes inside the elements.

Arguments

fstar1::AbstractArray{T} where T<:Real:

fstar2::AbstractArray{T} where T<:Real

fstar3::AbstractArray{T} where T<:Real

dg::Dg3D

u::AbstractArray{T} where T<:Real

element_id::Integer

Trixi.calcflux_fv! — Method

calcflux_fv!(fstar1, fstar2, u_leftright, u, element_id, dg::Dg2D)

Calculate the finite volume fluxes inside the elements.

Arguments

fstar1::AbstractArray{T} where T<:Real:

fstar2::AbstractArray{T} where T<:Real

dg::Dg2D

u::AbstractArray{T} where T<:Real

element_id::Integer

Trixi.compute_jacobian_dg — Method

compute_jacobian_dg(parameters_file=nothing; verbose=false, parameters...)

Uses DG right hand side operator and simple second order finite difference to compute the Jacobian J of the operator. The linearisation state is the initial condition from the parameter file. Returns J.

Trixi.compute_linear_structure — Function

compute_linear_structure(parameters_file, source_terms=nothing; verbose=false, parameters...)

Computes the exact Jacobian A of a linear DG operator wrapped as a LinearMap and the right hand side b. Returns A, b.

Trixi.convtest — Method

convtest(parameters_file, iterations; parameters...)

Run multiple Trixi simulations with the parameters in parameters_file and compute the experimental order of convergence (EOC) in the $L^2$ and $L^\infty$ norm. The number of runs is specified by iterations and in each run the initial refinement level will be increased by 1. Parameters can be overriden by specifying them as additional keyword arguments, which are passed to the respective call to run..

Trixi.default_example — Method

default_example()

Return the path to an example parameter file that can be used to quickly see Trixi in action.

Trixi.element_solutions_to_mortars! — Method

element_solutions_to_mortars!(dg::Dg2D, ::Val{:l2}, leftright, m, u_large)

Interpolate u_large to dg.l2mortars.u_[upper/lower] for mortar m using the forward mortar operators of dg.

Trixi.element_solutions_to_mortars! — Method

element_solutions_to_mortars!(dg::Dg3D, ::Val{:l2}, leftright, m, u_large, fstar_tmp1)

Interpolate u_large to dg.l2mortars.u_[upper/lower]_[left/right] for mortar m using the forward mortar operators of dg and fstar_tmp1 as temporary storage.

Trixi.examples_dir — Method

examples_dir()

Return the directory where the example files provided with Trixi.jl are located. If Trixi is installed as a regular package (with ]add Trixi), these files are read-only and should not be modified. To find out which files are available, use, e.g., readdir:

Examples

julia> readdir(examples_dir())
4-element Array{String,1}:
 "2d"
 "3d"
 "README.md"
 "paper-self-gravitating-gas-dynamics"

Trixi.flux_central — Method

flux_central(u_ll, u_rr, orientation, equation::AbstractEquation)

The classical central numerical flux f((u_ll) + f(u_rr)) / 2. When this flux is used as volume flux, the discretization is equivalent to the classical weak form DG method (except floating point errors).

Trixi.flux_chandrashekar — Method

flux_chandrashekar(u_ll, u_rr, orientation, equation::CompressibleEulerEquations2D)

Entropy conserving two-point flux by Chandrashekar (2013) Kinetic Energy Preserving and Entropy Stable Finite Volume Schemes for Compressible Euler and Navier-Stokes Equations DOI: 10.4208/cicp.170712.010313a

Trixi.flux_chandrashekar — Method

flux_chandrashekar(u_ll, u_rr, orientation, equation::CompressibleEulerEquations3D)

Entropy conserving two-point flux by Chandrashekar (2013) Kinetic Energy Preserving and Entropy Stable Finite Volume Schemes for Compressible Euler and Navier-Stokes Equations DOI: 10.4208/cicp.170712.010313a

Trixi.flux_derigs_etal — Method

flux_derigs_etal(u_ll, u_rr, orientation, equation::IdealGlmMhdEquations2D)

Entropy conserving two-point flux by Derigs et al. (2018) Ideal GLM-MHD: About the entropy consistent nine-wave magnetic field divergence diminishing ideal magnetohydrodynamics equations DOI: 10.1016/j.jcp.2018.03.002

Trixi.flux_derigs_etal — Method

flux_derigs_etal(u_ll, u_rr, orientation, equation::IdealGlmMhdEquations3D)

Entropy conserving two-point flux by Derigs et al. (2018) Ideal GLM-MHD: About the entropy consistent nine-wave magnetic field divergence diminishing ideal magnetohydrodynamics equations DOI: 10.1016/j.jcp.2018.03.002

Trixi.flux_hll — Method

flux_hll(u_ll, u_rr, orientation, equation::IdealGlmMhdEquations2D)

HLL flux for ideal GLM-MHD equations like that by Li (2005) An HLLC Riemann solver for magneto-hydrodynamics DOI: 10.1016/j.jcp.2004.08.020

Trixi.flux_hll — Method

flux_hll(u_ll, u_rr, orientation, equation::IdealGlmMhdEquations3D)

HLL flux for ideal GLM-MHD equations like that by Li (2005) An HLLC Riemann solver for magneto-hydrodynamics DOI: 10.1016/j.jcp.2004.08.020

Trixi.flux_hllc — Method

flux_hllc(u_ll, u_rr, orientation, equation::CompressibleEulerEquations3D)

Computes the HLLC flux (HLL with Contact) for compressible Euler equations developed by E.F. Toro http://www.prague-sum.com/download/2012/Toro_2-HLLC-RiemannSolver.pdf Signal speeds: https://sci-hub.tw/https://doi.org/10.1137/S1064827593260140

Trixi.flux_kennedy_gruber — Method

flux_kennedy_gruber(u_ll, u_rr, orientation, equation::CompressibleEulerEquations2D)

Kinetic energy preserving two-point flux by Kennedy and Gruber (2008) Reduced aliasing formulations of the convective terms within the Navier-Stokes equations for a compressible fluid DOI: 10.1016/j.jcp.2007.09.020

Trixi.flux_kennedy_gruber — Method

flux_kennedy_gruber(u_ll, u_rr, orientation, equation::CompressibleEulerEquations3D)

Kinetic energy preserving two-point flux by Kennedy and Gruber (2008) Reduced aliasing formulations of the convective terms within the Navier-Stokes equations for a compressible fluid DOI: 10.1016/j.jcp.2007.09.020

Trixi.flux_ranocha — Method

flux_ranocha(u_ll, u_rr, orientation, equation::CompressibleEulerEquations2D)

Entropy conserving and kinetic energy preserving two-point flux by Ranocha (2018) Generalised Summation-by-Parts Operators and Entropy Stability of Numerical Methods for Hyperbolic Balance Laws PhD thesis, TU Braunschweig See also Ranocha (2020) Entropy Conserving and Kinetic Energy Preserving Numerical Methods for the Euler Equations Using Summation-by-Parts Operators Proceedings of ICOSAHOM 2018

Trixi.flux_ranocha — Method

flux_ranocha(u_ll, u_rr, orientation, equation::CompressibleEulerEquations3D)

Entropy conserving and kinetic energy preserving two-point flux by Ranocha (2018) Generalised Summation-by-Parts Operators and Entropy Stability of Numerical Methods for Hyperbolic Balance Laws PhD thesis, TU Braunschweig See also Ranocha (2020) Entropy Conserving and Kinetic Energy Preserving Numerical Methods for the Euler Equations Using Summation-by-Parts Operators Proceedings of ICOSAHOM 2018

Trixi.flux_shima_etal — Method

function flux_shima_etal(u_ll, u_rr, orientation, equation::CompressibleEulerEquations2D)

This flux is is a modification of the original kinetic energy preserving two-point flux by Kuya, Totani and Kawai (2018) Kinetic energy and entropy preserving schemes for compressible flows by split convective forms DOI: 10.1016/j.jcp.2018.08.058 The modification is in the energy flux to guarantee pressure equilibrium and was developed by Nao Shima, Yuichi Kuya, Yoshiharu Tamaki, Soshi Kawai (JCP 2020) Preventing spurious pressure oscillations in split convective form discretizations for compressible flows

Trixi.flux_shima_etal — Method

function flux_shima_etal(u_ll, u_rr, orientation, equation::CompressibleEulerEquations3D)

This flux is is a modification of the original kinetic energy preserving two-point flux by Kuya, Totani and Kawai (2018) Kinetic energy and entropy preserving schemes for compressible flows by split convective forms DOI: 10.1016/j.jcp.2018.08.058 The modification is in the energy flux to guarantee pressure equilibrium and was developed by Nao Shima, Yuichi Kuya, Yoshiharu Tamaki, Soshi Kawai (JCP 2020) Preventing spurious pressure oscillations in split convective form discretizations for compressible flows

Trixi.get_examples — Method

get_examples()

Return a list of all example parameter files that are provided by Trixi.

Trixi.get_name — Method

get_name(x)

Returns a name of x ready for pretty printing. By default, return string(y) if x isa Val{y} and return string(x) otherwise.

Examples

julia> Trixi.get_name("test")
"test"

julia> Trixi.get_name(Val(:test))
"test"

Trixi.get_node_coords — Method

get_node_coords(x, dg::AbstractDg, indices...)

Return an ndims(dg)-dimensional SVector for the DG node specified via the i, j, k, element_id indices (3D) or i, j, element_id indices (2D).

Trixi.get_node_vars — Method

get_node_vars(u, dg::AbstractDg, indices...)

Return an nvariables(dg)-dimensional SVector of the conservative variables for the DG node specified via the i, j, k, element_id indices (3D) or i, j, element_id indices (2D).

Trixi.initial_conditions_density_wave — Method

initial_conditions_density_wave(x, t, equation::CompressibleEulerEquations2D)

test case for stability of EC fluxes from paper: https://arxiv.org/pdf/2007.09026.pdf domain [-1, 1]^2 mesh = 4x4, polydeg = 5

Trixi.integrate — Method

integrate(func, u, dg::Dg2D; normalize=true)
integrate(u, dg::Dg2D; normalize=true)

Call function func for each DG node and integrate the result over the computational domain.

The function func is called as func(u_local) for each volume node (i, j) and each element_id, where u_local is an SVectorized copy of u[:, i, j, element_id]. If normalize is true, the result is divided by the total volume of the computational domain. If func is omitted, it defaults to identity.

Examples

Calculate the integral over all conservative variables:

state_integrals = integrate(dg.elements.u, dg)

Trixi.integrate — Method

integrate(func, u, dg::Dg3D; normalize=true)
integrate(u, dg::Dg3D; normalize=true)

Call function func for each DG node and integrate the result over the computational domain.

The function func is called as func(u_local) for each volume node (i, j, k) and each element_id, where u_local is an SVectorized copy of u[:, i, j, k, element_id]. If normalize is true, the result is divided by the total volume of the computational domain. If func is omitted, it defaults to identity.

Examples

Calculate the integral over all conservative variables:

state_integrals = integrate(dg.elements.u, dg)

Trixi.integrate — Method

integrate(func, dg::Dg2D, args...; normalize=true)

Call function func for each DG node and integrate the result over the computational domain.

The function func is called as func(i, j, element_id, dg, args...) for each volume node (i, j) and each element_id. Additional positional arguments args... are passed along as well. If normalize is true, the result is divided by the total volume of the computational domain.

Examples

Calculate the integral of the time derivative of the entropy, i.e., ∫(∂S/∂t)dΩ = ∫(∂S/∂u ⋅ ∂u/∂t)dΩ:

# Calculate integral of entropy time derivative
dsdu_ut = integrate(dg, dg.elements.u, dg.elements.u_t) do i, j, element_id, dg, u, u_t
  u_node   = get_node_vars(u,   dg, i, j, element_id)
  u_t_node = get_node_vars(u_t, dg, i, j, element_id)
  dot(cons2entropy(u_node, equations(dg)), u_t_node)
end

Trixi.integrate — Method

integrate(func, dg::Dg3D, args...; normalize=true)

Call function func for each DG node and integrate the result over the computational domain.

The function func is called as func(i, j, k, element_id, dg, args...) for each volume node (i, j, k) and each element_id. Additional positional arguments args... are passed along as well. If normalize is true, the result is divided by the total volume of the computational domain.

Examples

Calculate the integral of the time derivative of the entropy, i.e., ∫(∂S/∂t)dΩ = ∫(∂S/∂u ⋅ ∂u/∂t)dΩ:

# Calculate integral of entropy time derivative
dsdu_ut = integrate(dg, dg.elements.u, dg.elements.u_t) do i, j, k, element_id, dg, u, u_t
  u_node   = get_node_vars(u,   dg, i, j, k, element_id)
  u_t_node = get_node_vars(u_t, dg, i, j, k, element_id)
  dot(cons2entropy(u_node, equations(dg)), u_t_node)
end

Trixi.mortar_fluxes_to_elements! — Method

mortar_fluxes_to_elements!(surface_flux_values, dg::Dg2D, mortar_type::Val{:l2}, m, fstar_upper, fstar_lower)

Copy/project fstar_[upper/lower] to surface_flux_values for mortar m using the reverse mortar operators of dg.

Trixi.mortar_fluxes_to_elements! — Method

mortar_fluxes_to_elements!(surface_flux_values, dg::Dg3D, mortar_type::Val{:l2}, m,
                           fstar_upper_left, fstar_upper_right,
                           fstar_lower_left, fstar_lower_right,
                           fstar_tmp1)

Copy/project fstar_[upper/lower]_[left/right to surface_flux_values for mortar m using the reverse mortar operators of dg and fstar_tmp1 as temporary storage.

Trixi.multiply_dimensionwise — Method

multiply_dimensionwise(matrix::AbstractMatrix, data_in::AbstractArray{<:Any, NDIMS+1})

Multiply the array data_in by matrix in each coordinate direction, where data_in is assumed to have the first coordinate for the number of variables and the remaining coordinates are multiplied by matrix.

Trixi.pure_and_blended_element_ids! — Method

pure_and_blended_element_ids!(element_ids_dg, element_ids_dgfv, alpha, dg)

Given blending factors alpha and the solver dg, fill element_ids_dg with the IDs of elements using a pure DG scheme and element_ids_dgfv with the IDs of elements using a blended DG-FV scheme.

Trixi.pure_and_blended_element_ids! — Method

pure_and_blended_element_ids!(element_ids_dg, element_ids_dgfv, alpha, dg)

Given blending factors alpha and the solver dg, fill element_ids_dg with the IDs of elements using a pure DG scheme and element_ids_dgfv with the IDs of elements using a blended DG-FV scheme.

Trixi.run — Method

run(parameters_file; verbose=false, refinement_level_increment=0, parameters...)

Run a Trixi simulation with the parameters in parameters_file. Parameters can be overriden by specifying them as keyword arguments (see examples).

If verbose is true, additional output will be generated on the terminal that may help with debugging. If a value for refinement_level_increment is given, initial_refinement_level will be increased by this value before running the simulation (mostly used by EOC analysis).

Examples

julia> Trixi.run("examples/parameters.toml", verbose=true)
[...]

Without changing the parameters file we can start a simulation with polydeg = 1 and t_end = 0.5 as follows:

julia> Trixi.run("examples/parameters.toml", polydeg=1, t_end=0.5)
[...]

Trixi.save_analysis_header — Method

save_analysis_header(filename, quantities, equation)

Truncate file filename and save a header with the names of the quantities quantities that will subsequently written to filename by analyze_solution. Since some quantities are equation-specific, the system of equations instance is passed in equation.

Note: Keep order of analysis quantities in sync with analyze_solution when adding or changing quantities.

Trixi.save_analysis_header — Method

save_analysis_header(filename, quantities, equation)

Truncate file filename and save a header with the names of the quantities quantities that will subsequently written to filename by analyze_solution. Since some quantities are equation-specific, the system of equations instance is passed in equation.

Note: Keep order of analysis quantities in sync with analyze_solution when adding or changing quantities.

Trixi.setparameter — Method

setparameter(name::String, value)

Set parameter with the specified name to the specified value.

Trixi.timestep_2N! — Method

timestep_2N!(solver, t, dt, a, b, c)

Perform one timestep using an explicit Runge-Kutta method of the low-storage class 2N of Williamson.

Trixi.timestep_3Sstar! — Method

timestep_3Sstar!(solver, t, dt, gamma1, gamma2, gamma3, beta, delta, c)

Perform one timestep using an explicit Runge-Kutta method of the low-storage class 3Sstar of Ketcheson.

Trixi.timestep_carpenter_kennedy_erk43_2N! — Method

timestep_carpenter_kennedy_erk43_2N!(solver::AbstractSolver, t, dt)

Carpenter, Kennedy (1994) Third order 2N storage RK schemes with error control

Trixi.timestep_carpenter_kennedy_erk54_2N! — Method

timestep_carpenter_kennedy_erk54_2N!(solver::AbstractSolver, t, dt)

Carpenter, Kennedy (1994) Fourth order 2N storage RK schemes, Solution 3

Trixi.timestep_hyp_diff_N3_erk52_3Sstar! — Method

timestep_hyp_diff_N3_erk52_3Sstar!(solver::AbstractSolver, t, dt)

Five stage, second-order acurate explicit Runge-Kutta scheme with stability region optimized for the hyperbolic diffusion equation with LLF flux and polynomials of degree polydeg=3.

Trixi.timestep_parsani_ketcheson_deconinck_erk32_3Sstar! — Method

timestep_parsani_ketcheson_deconinck_erk32_3Sstar!(solver::AbstractSolver, t, dt)

Parsani, Ketcheson, Deconinck (2013) Optimized explicit RK schemes for the spectral difference method applied to wave propagation problems DOI: 10.1137/120885899

Trixi.timestep_parsani_ketcheson_deconinck_erk94_3Sstar! — Method

timestep_parsani_ketcheson_deconinck_erk94_3Sstar!(solver::AbstractSolver, t, dt)

Parsani, Ketcheson, Deconinck (2013) Optimized explicit RK schemes for the spectral difference method applied to wave propagation problems DOI: 10.1137/120885899